Refactor code generation to use Blocks architecture
- Separated functions from functions.ts into smaller, focused modules
- Introduced a new Blocks-based approach for code generation
- Restructured the project directory:
  * Added generator/blocks directory for block-specific logic
  * Created block-generator.ts for handling block-to-TypeScript conversion
  * Updated code-generator.ts to use the new Blocks system
- Migrated existing functionality to work with the new Blocks structure
- Temporarily removed unit tests (will be added later)
