+++ packages/ts-generator/src/types/blocks.ts
@@ -1,16 +1,17 @@
-import { VariableInfo } from "./common";
+import { FunctionInfo, VariableInfo } from "./common";
 
 export interface Block {
   index: number;
   comment?: string;
+  blockType: string;
 }
 
 export interface FunctionCallBlock extends Block {
-  type: string;
-  functionName: string;
+  functionInfo: FunctionInfo;
   parameters?: VariableInfo[];
   returnVariable?: VariableInfo;
   isAsync: boolean;
+  blockType: "functionCall";
 }
 
 // Add more block types as implemented
